+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 
 - Add support for multiple plone sites on one installation via PLONE_BACKEND_HOST @maethu
 
+- [Issue #118](https://github.com/collective/collective.elasticsearch/issues/118) Fix **ComponentLookupError** when adding a Plone Site (6.1) (@andreclimaco)
+
 
 ## 5.0.0 (2022-10-11)